requirements.txt
pandas~=2.0.1
matplotlib~=3.7.1
scikit-learn~=1.2.2
spacy~=3.5.2
numpy~=1.24.3
wordcloud==1.9.1.1
Pillow~=9.5.0
pypdf2~=3.0.1
Django==4.2.1